pyproject.toml
[tool.poetry]
name = "describebot"
version = "0.1.0"
description = ""
authors = ["ashishdatta <ashishbdatta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
replicate = "^0.15.4"
pillow = "^10.1.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
